refactor: 모델 패키지 분리
