+++ src/main/java/roomescape/dto/CreateReservationRequest.java
@@ -2,7 +2,7 @@ package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.Reservation;
+import roomescape.model.Reservation;
 
 public record CreateReservationRequest(String name, LocalDate date, LocalTime time) {
